test: added test for service :white_check_mark:
