+++ src/main/java/fr/octocorn/happyningbackend/event/exception/BadRequestDto.java
@@ -1,0 +1,5 @@
+package fr.octocorn.happyningbackend.event.exception;
+
+public class BadRequestDto {
+
+}
