+++ src/main/java/fr/octocorn/happyningbackend/event/exception/BadRequestException.java
@@ -8,4 +8,5 @@ public class BadRequestException extends RuntimeException {
     public BadRequestException(String message) {
         super(message);
     }
+
 }
